+++ cmd/frontend/internal/app/jscontext/jscontext.go
@@ -65,6 +65,7 @@ type JSContext struct {
 	// Experimental features
 	SearchTimeoutParameterEnabled bool `json:"searchTimeoutParameterEnabled"`
 	ShowMissingReposEnabled       bool `json:"showMissingReposEnabled"`
+	HostSurveysLocallyEnabled     bool `json:"hostSurveysLocallyEnabled"`
 
 	AccessTokensEnabled bool `json:"accessTokensEnabled"`
 }
@@ -140,6 +141,7 @@ func NewJSContextFromRequest(req *http.Request) JSContext {
 		// do the default behavior only in Go land.
 		SearchTimeoutParameterEnabled: conf.SearchTimeoutParameterEnabled(),
 		ShowMissingReposEnabled:       conf.ShowMissingReposEnabled(),
+		HostSurveysLocallyEnabled:     conf.HostSurveysLocallyEnabled(),
 
 		AccessTokensEnabled: conf.AccessTokensEnabled(),
 	}