+++ pkg/hubspot/hubspot.go
@@ -5,11 +5,14 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"io/ioutil"
 	"net/http"
 	"net/url"
 	"path"
+	"strings"
 	"time"
 
+	"github.com/google/go-querystring/query"
 	"golang.org/x/net/context/ctxhttp"
 )
 
@@ -25,6 +28,47 @@ func New(portalID string, hapiKey string) *Client {
 		portalID: portalID,
 		hapiKey:  hapiKey,
 	}
+}
+
+// Send a POST request with form data to HubSpot APIs that accept
+// application/x-www-form-urlencoded data (e.g. the Forms API)
+func (c *Client) postForm(methodName string, baseURL *url.URL, suffix string, body interface{}) error {
+	var data url.Values
+	switch body := body.(type) {
+	case map[string]string:
+		data = make(url.Values, len(body))
+		for k, v := range body {
+			data.Set(k, v)
+		}
+	default:
+		var err error
+		data, err = query.Values(body)
+		if err != nil {
+			return wrapError(methodName, err)
+		}
+	}
+
+	baseURL.Path = path.Join(baseURL.Path, suffix)
+	req, err := http.NewRequest("POST", baseURL.String(), strings.NewReader(data.Encode()))
+	if err != nil {
+		return wrapError(methodName, err)
+	}
+	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return wrapError(methodName, err)
+	}
+	defer resp.Body.Close()
+	if resp.StatusCode != http.StatusNoContent && resp.StatusCode != http.StatusFound {
+		buf, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			return wrapError(methodName, err)
+		}
+		return wrapError(methodName, fmt.Errorf("Code %v: %s", resp.StatusCode, string(buf)))
+	}
+
+	return nil
 }
 
 // Send a POST request with JSON data to HubSpot APIs that accept JSON